Revert "feat!(themes): structured everything for adding new themes"
This reverts commit a769aea65bebdc6288f477a80cb0bd2218ab88c1.
