+++ src/settings/settings.ts
@@ -1,5 +1,3 @@
-import { ThemeStyle } from './themes';
-
 export class DiscordRPCSettings {
   showVaultName: boolean = true;
   showCurrentFileName: boolean = true;
@@ -19,4 +17,19 @@ export enum PluginState {
   connected,
   connecting,
   disconnected,
+}
+
+export enum ThemeStyle {
+  Default_dark = "default-old-dark",
+  Default_light = "default-old-light",
+  Default_new_dark = "default-new-dark",
+  Default_new_light = "default-new-light",
+  Catppuccin_Latte = "latte",
+  Catppuccin_Frappe = "frappe",
+  Catppuccin_Macchiato = "macchiato",
+  Catppuccin_Mocha = "mocha",
+  Cyberglow_Dark = "cyberglow-dark",
+  Cyberglow_Light = "cyberglow-light",
+  Tokyo_night_Dark = "tokyo-night-dark",
+  Tokyo_night_Light = "tokyo-night-light"
 }
